Add person's model

+++ src/main/java/com/example/spring/demo/service/PersonService.java
@@ -2,9 +2,12 @@ package com.example.spring.demo.service;
 
 
 import com.example.spring.demo.entity.PersonEntity;
+import com.example.spring.demo.model.Person;
 import com.example.spring.demo.repository.PersonRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
 
 @Service
 public class PersonService {
@@ -14,34 +17,39 @@ public class PersonService {
         this.personRepository = personRepository;
     }
 
-    public PersonEntity registration(PersonEntity person) {
-        return personRepository.save(person);
+    public Person registration(PersonEntity person) {
+        return Person.toModel(personRepository.save(person));
     }
 
-    public Iterable<PersonEntity> findAll() {
-        return personRepository.findAll();
+    public ArrayList<Person> findAll() {
+        ArrayList<Person> people = new ArrayList<>();
+
+        for (PersonEntity entity : personRepository.findAll()) {
+            people.add(Person.toModel(entity));
+        }
+
+        return people;
     }
 
-    public PersonEntity findById(Integer id) {
-        return personRepository.findById(id).get();
+    public Person findById(Integer id) {
+        return Person.toModel(personRepository.findById(id).get());
     }
 
-    public PersonEntity findByName(String name) {
-        return personRepository.findByName(name);
+    public Person findByName(String name) {
+        return Person.toModel(personRepository.findByName(name));
     }
 
     public Integer deleteById(Integer id) {
-        personRepository.deleteById(id);
-        return id;
+        if (personRepository.existsById(id)) {
+            personRepository.deleteById(id);
+            return 0;
+        }
+        return -1;
     }
 
     public Integer deleteAll() {
         personRepository.deleteAll();
         return 0;
-    }
-
-    public boolean existById(Integer id) {
-        return personRepository.existsById(id);
     }
 
 }
